pyproject.toml
[project]
name = "django_chat"
version = "1.0.1"
description = "chat app"
authors = [
    {name = "liaozhiming", email = "liaozhimingandy@qq.com"},
]
dependencies = [
    "django>=5.0.3",
    "gunicorn>=21.2.0",
    "docutils>=0.20.1",
    "django-cors-headers>=4.3.1",
    "django-filter>=24.2",
    "Pillow>=10.2.0",
    "pyjwt>=2.8.0",
    "minio>=7.2.5",
    "pymssql>=2.3.0",
    "psycopg2-binary>=2.9.9",
    "django-ratelimit>=4.1.0",
    "djangorestframework>=3.15.1",
    "markdown>=3.6",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "Apache-2.0"}

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
]
